ml-engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
mlxtend==0.23.1
pymongo==4.6.1
python-dotenv==1.0.1
# uvicorn app:app --reload --port 8000
